+++ dev/src/components/profile/ProfileForm.tsx
@@ -1,7 +1,7 @@
 import React, { Dispatch, SetStateAction, useMemo } from "react"
 import { useActions, useAppState } from "../../overmind"
 import { json } from "overmind"
-import FormInput from "./FormInput"
+import FormInput from "../forms/FormInput"
 
 export const ProfileForm = ({ children, setEditing }: { children: React.ReactNode, setEditing: Dispatch<SetStateAction<boolean>> }): JSX.Element => {
     const state = useAppState()
